ImageForm: refactoring directories, add disabling button Cropper: add avatarMode
